Move ExperienceForms project to /code folder
See http://helix.sitecore.net/principles/file-structure/modules.html
