+++ Features/ExperienceForms/code/Model/SendMail/SendMailActionData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Stockpick.Forms.Feature.ExperienceForms.Model.SendMail
+{
+    public class SendMailActionData
+    {
+        public Guid ReferenceId { get; set; }
+    }
+}
