--- Features/ExperienceForms/code/Stockpick.Forms.Feature.ExperienceForms/Model/SendMail/SendEmailTemplates.scx
@@ -1,18 +1,0 @@
-﻿<CodeGeneration Generator="Template Class">
-  <shape designerX="11" designerY="12">
-    <template templatename="E-mail Templates Folder" templateid="{EF52589A-C7CD-413A-9B29-8DC8D8146EC2}" templateuri="/sc9u1.local/master/{EF52589A-C7CD-413A-9B29-8DC8D8146EC2}" classname="EmailTemplatesFolder" />
-  </shape>
-  <shape designerX="11" designerY="58">
-    <template templatename="E-mail Template" templateid="{1FF3E9C9-73D7-4395-960A-7271EED8CC1E}" templateuri="/sc9u1.local/master/{1FF3E9C9-73D7-4395-960A-7271EED8CC1E}" classname="EmailTemplate">
-      <section id="{82B0D786-3222-4E8C-904D-A86FD73B97EE}" name="E-mail Setup">
-        <field id="{BD289C06-A59E-4F38-B824-268973F3EDA9}" name="Subject" shared="0" source="" type="Single-Line Text" unversioned="0" propertyname="Subject" />
-        <field id="{DF91B8A3-7262-42EA-96F0-74796A0A9BC7}" name="From" shared="0" source="" type="Single-Line Text" unversioned="0" propertyname="From" />
-        <field id="{95FC9607-A6E1-40E2-B4DD-7B7D23E97DEC}" name="To" shared="0" source="" type="Single-Line Text" unversioned="0" propertyname="To" />
-        <field id="{A5C64FE9-A29E-43CB-9861-D696F31BC586}" name="Cc" shared="0" source="" type="Single-Line Text" unversioned="0" propertyname="Cc" />
-        <field id="{BB6A889B-CF26-4188-9882-5498C9E64CAA}" name="Bcc" shared="0" source="" type="Single-Line Text" unversioned="0" propertyname="Bcc" />
-        <field id="{5EAF34E9-F237-481D-A20F-9704EE7B88D7}" name="Message RichText" shared="0" source="site:Richtext Editor Profile" type="Rich Text" unversioned="0" propertyname="MessageRichText" />
-        <field id="{4157F17D-CDA9-4CFE-AF09-F472635C87ED}" name="Message Text" shared="0" source="" type="Multi-Line Text" unversioned="0" propertyname="MessageText" />
-      </section>
-    </template>
-  </shape>
-</CodeGeneration>
