--- Features/ExperienceForms/code/Stockpick.Forms.Feature.ExperienceForms/Model/SendMail/SendMailActionData.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Stockpick.Forms.Feature.ExperienceForms.Model.SendMail
-{
-    public class SendMailActionData
-    {
-        public Guid ReferenceId { get; set; }
-    }
-}
